refactor(runtime): canonicalize public runtime API surface
Prune the pre-alpha runtime API surface down to canonical public names and
check in the runtime API surface review that drove the cleanup. Remove public
RT_ALIAS support from rtgen, make RT_ALIAS a generator error, stop alias rows
from feeding the name map/signature/frontend-name outputs, and prevent
functions in descendant namespaces from being synthesized as methods on their
parent runtime classes. Make --dump-runtime-api report the public signature text
instead of the lowered descriptor form so typed object signatures stay visible
in the API inventory.

Normalize the remaining runtime surface around the intended names. Keep
canonical Core.Box option conversions, Core.Parse Try* helpers, collection To*
conversion direction, Length cardinality properties, Path-suffixed canvas path
operations, and the corrected Game3D/Graphics3D class leaves. Refresh the
surface policy, runtime name/signature helpers, ownership metadata, source
health baseline, generated docs/site pages, release notes, and runtime catalog
coverage for the smaller surface.

Migrate the tree to the canonical API names across BASIC and Zia examples,
ViperIDE, runtime sweeps, conformance probes, graphics3d demos, app demos, and
golden IL expectations. Update tests that asserted duplicate or alias-heavy
behavior so they now guard the canonical surface instead, including the runtime
class qualified surface, wishlist bindings, graphics3d ABI surface, BASIC
runtime calls, collection/parse/box runtime tests, and Zia static/runtime call
coverage.

Repair cleanup fallout in shipped demos by moving MusicGen.SetLength call sites
to the Length property and replacing removed unqualified Int formatting calls
with Viper.Text.Fmt.Int. Add the macOS VIPER_GFX_NO_ACTIVATE graphics knob,
apply it to requires_display CTest cases, and document it so graphics tests can
run without repeatedly activating their windows.

Tests: Not run per user request.

+++ examples/apiaudit/core/convert_demo.bas
@@ -15,9 +15,9 @@ PRINT "ToDouble('-2.5'): "; Viper.Core.Convert.ToDouble("-2.5")
 
 ' --- Convert.ToInt ---
 PRINT "--- Convert.ToInt ---"
-PRINT "ToInt('42'): "; Viper.Core.Convert.ToInt("42")
-PRINT "ToInt('0'): "; Viper.Core.Convert.ToInt("0")
-PRINT "ToInt('-100'): "; Viper.Core.Convert.ToInt("-100")
+PRINT "ToInt('42'): "; Viper.Core.Convert.ToInt64("42")
+PRINT "ToInt('0'): "; Viper.Core.Convert.ToInt64("0")
+PRINT "ToInt('-100'): "; Viper.Core.Convert.ToInt64("-100")
 
 ' --- Convert.ToInt64 ---
 PRINT "--- Convert.ToInt64 ---"
