+++ src/tests/golden/basic_runtime_ns/convert_to_int.bas
@@ -1,3 +1,3 @@
-REM BASIC: Convert string to int via Viper.Core.Convert.ToInt
+REM BASIC: Convert string to int via Viper.Core.Convert.ToInt64
 USING Viper.Terminal
-Viper.Terminal.PrintI64(Viper.Core.Convert.ToInt("42"))
+Viper.Terminal.PrintI64(Viper.Core.Convert.ToInt64("42"))
